trade/requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
openpyxl>=3.1.2

# For enhanced features (optional)
# neo4j>=5.14.0  # For ICIJ graph database
# beautifulsoup4>=4.12.0  # For web scraping
# comtradeapicall>=0.2.0  # For UN Comtrade (if available)
# streamlit>=1.28.0  # For web UI (future)
